backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.1
python-dotenv==1.0.1
bcrypt==4.1.3

SQLAlchemy==2.0.30
alembic==1.13.2
